Add projectsornamespacesfilter
Use informer for projectsornamespaces filter

Prior, a NamespaceCache needed to be passed to filter namespaces
belonging to a project. Now, a sqlcache informer is used to
query which namespaces either share a name with a value in the
projectsornamespaces query parameter or belong to a project that
shares a name with one of the values.